Move Builder as non inner class in prometheus exporter
